Specify exact license in header of Emacs lisp library (#353)
* Emacs: specify exact license in library header

Emacs packages are often distributed by themselves (e.g. on Melpa
and the Emacsmirror) even when they are developed in a repository
that other files beside the Elisp libraries.  In such cases it
should still be possible to tell what the license is, so replace
"BSD-style license" with "BSD-3-clause" and use the standard
"License" header keyword to do so.

* Emacs: cleanup whitespace

* Emacs: golint: wrap and slightly improve doc-string